Fixed TopSpeakerController Issues

+++ src/api/controllers/TopSpeaker.controller.ts
@@ -2,15 +2,13 @@
 import { Express, Request, Response, NextFunction } from "express";
 import TopSpeakerService from "../services";
 import logger from "../../util/logger";
+import { ITopSpeaker } from "../interfaces";
 
 /**
- * @todo implement a @function insertTopSpeaker that calls 
- * @function insertTopSpeaker in the TopSpeakerService 
- * 
  * @param {Request} request - Request from the frontend
  * @param {Response} response - Response that need to send to the client
  * @param {NextFunction} next - Next function
- * @returns inserted topSpeaker
+ * @returns { ITopSpeaker } topSpeaker document
  */
 
  export const insertTopSpeaker = async (
@@ -30,13 +28,10 @@ import logger from "../../util/logger";
   };
 
 /**
- * @todo implement a @function getTopSpeaker that calls 
- * @function getTopSpeaker in the TopSpeakerService 
- * 
  * @param {Request} request - Request from the frontend
  * @param {Response} response - Response that need to send to the client
  * @param {NextFunction} next - Next function
- * @returns topSpeaker
+ * @returns { ITopSpeaker } topSpeaker document
  */
 
  export const getTopSpeaker = async (
@@ -44,25 +39,25 @@ import logger from "../../util/logger";
     response: Response,
     next: NextFunction
   ) => {
-    await TopSpeakerService.getTopSpeaker(request.params.topSpeakerId)
+    let topSpeakerId = request.params.topSpeakerId;
+
+    if (topSpeakerId) {
+      await TopSpeakerService.getTopSpeaker(request.params.topSpeakerId)
       .then((data) => {
         request.handleResponse.successRespond(response)(data);
         next();
       })
-      .catch((error: any) => {
-        request.handleResponse.errorRespond(response)(error.message);
-        next();
-      });
-  };
+    }else{
+        request.handleResponse.errorRespond(response)('Top Speaker ID not found');
+      };
+    };
+  
 
 /**
- * @todo implement a @function getTopSpeakers that calls 
- * @function getTopSpeakers in the TopSpeakerService 
- * 
  * @param {Request} request - Request from the frontend
  * @param {Response} response - Response that need to send to the client
  * @param {NextFunction} next - Next function
- * @returns topSpeaker[]
+ * @returns { ITopSpeaker[] } All top speakers in the system
  */
 
  export const getTopSpeakers = async (
@@ -83,13 +78,10 @@ import logger from "../../util/logger";
 
 
 /**
- * @todo implement a @function updateTopSpeaker that calls 
- * @function updateTopSpeaker in the TopSpeakerService 
- * 
  * @param {Request} request - Request from the frontend
  * @param {Response} response - Response that need to send to the client
  * @param {NextFunction} next - Next function
- * @returns updated topSpeaker
+ * @returns { ITopSpeaker } - Updated top speaker details
  */
 
  export const updateTopSpeaker = async (
@@ -97,25 +89,25 @@ import logger from "../../util/logger";
     response: Response,
     next: NextFunction
   ) => {
-    await TopSpeakerService.updateTopSpeaker(request.params.topSpeakerId, request.body)
+
+    let topSpeakerId = request.params.topSpeakerId;
+
+    if (topSpeakerId) { 
+      await TopSpeakerService.updateTopSpeaker(request.params.topSpeakerId, request.body)
       .then((data) => {
         request.handleResponse.successRespond(response)(data);
         next();
       })
-      .catch((error: any) => {
-        request.handleResponse.errorRespond(response)(error.message);
-        next();
-      });
+    } else { 
+      request.handleResponse.errorRespond(response)('Top Speaker ID not found');
+    };
   };
 
 /**
- * @todo implement a @function deleteTopSpeaker that calls 
- * @function deleteTopSpeaker in the TopSpeakerService 
- * 
  * @param {Request} request - Request from the frontend
  * @param {Response} response - Response that need to send to the client
  * @param {NextFunction} next - Next function
- * @returns updated topSpeaker
+ * @returns { ITopSpeaker } - Deleted top speaker details
  */
 
  export const deleteTopSpeaker = async (
@@ -123,13 +115,17 @@ import logger from "../../util/logger";
     response: Response,
     next: NextFunction
   ) => {
-    await TopSpeakerService.deleteTopSpeaker(request.params.topSpeakerId)
+
+    let topSpeakerId = request.params.topSpeakerId;
+
+    if (topSpeakerId) {
+      await TopSpeakerService.deleteTopSpeaker(request.params.topSpeakerId)
       .then((data) => {
         request.handleResponse.successRespond(response)(data);
         next();
       })
-      .catch((error: any) => {
-        request.handleResponse.errorRespond(response)(error.message);
-        next();
-      });
+   } else {
+      request.handleResponse.errorRespond(response)('Top Speaker ID not found');
+   };
+    
   };
